backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
openai>=1.12.0
python-dotenv>=1.0.1
mangum>=0.17.0

sqlalchemy>=2.0.0
pyodbc>=5.0.0
